Disable WiFi, enable WiFi.
Turns out that, on wake, WiFi being enabled reliably means WiFi state is "NA" on wake. Unless we find a reliable way to boot the framework into a higher state of consciousness from RTC wake, the best option seems to be to enable WiFi on wake (actually connects WiFi that way), then disable when done. If we ever need WiFi in user-land, we can enable it.